Top-level hidden interface cleanup
- Deprecated top-level `result_from_sys`. Replaced by a hidden method.
- Made unreleased `report_init_error` private.
- Made `GODOT_API` and `GDNATIVE_LIBRARY_SYS` private.
- Moved `get_api`, `get_gdnative_library_sys` and `cleanup_internal_state`
  under `private`.
- Moved API initialization from `godot_gdnative_init` macro to
  `private::bind_api`.

Part of #357
