+++ gdnative-core/src/color_array.rs
@@ -1,5 +1,5 @@
 use crate::access::{Aligned, MaybeUnaligned};
-use crate::get_api;
+use crate::private::get_api;
 use crate::sys;
 use crate::Color;
 use crate::VariantArray;
