+++ gdnative-core/src/generated.rs
@@ -2,7 +2,7 @@
 #![allow(unused_imports)]
 
 use super::*;
-use crate::get_api;
+use crate::private::get_api;
 use crate::sys;
 use crate::sys::GodotApi;
